requirements.txt
Flask==3.0.0
groq==0.4.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-jose==3.3.0
bcrypt==4.1.2
